backend_crystal/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-image==0.22.0
openai==1.3.5
anthropic==0.7.7
google-generativeai==0.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4